Fixed race condition
It's possible that a packet is being processed when
poll() is called, but that it has not been enqueued yet.
If poll gets control, it checks the queue and then sleeps
if it can't find it's packet.  But if the check to notify
poll has already taken place, the intended packet is
enqueued after poll has checked for it.  Fix this by
extending the period that the packet processing routine
holds the mutex.