Added per-test priority support
In each module, a variable test_prio[] may be added which
is indexed by the test name.

Currently the only application of this is to exclude tests
from the "default" set of tests that are run:  If a test
has a negative priority, then it is not run by default.  This
addresses the issue with "fill exact match table" taking almost
an hour to run.