Added switch profile command option
Added --profile command line argument.  If present, it must specify
a python file which is imported.  Currently the only profile related
code is a variable called skip_test_list which gives a list of tests
that should _not_ be run for the platform.  Only the test names
are currently checked; test-module checking is not checked.

A sample profile file is also included.