Added barrier under load test, LoadBarrier
This test is still...in testing.  It assumes a loopback port is
configured (which can be specified with lb_port, defaults to 1).
A flow is inserted to create a storm of packet in messages from
the loopback.  Then a number of barrier messages are called.
The test passes if all barrier replies are received.  After the
barrier replies are received, the flow table is cleared.

The test is not run by default.