OFTEST-7 Add test for a bad message
Note you may need to run make install in tools/munger to pick
up the new illegal_message module before this test can be run.

Generate a message with an illegal type and verify that the
switch replies with an error message with the same XID and the
proper type and code.