speed up DirectBadPacketBase tests
These tests send many packets that are not expected to match, which is slow
because each non-matching packet is waited on for the full default timeout.
This change speeds up these tests in two ways:

1. Packets expected to not match have a shorter timeout.

2. A low priority "catch-all" flow is installed. Packets not dropped by the
switch should match this flow and be returned by dataplane.poll on the wrong
port.

DirectBadIpTcpPackets now runs in 10s instead of 140s.