pktact: remove exact match flows from wildcard priority tests
The OF 1.0 spec states that "An entry that specifies an exact match (i.e., it
has no wildcards) is always the highest priority." The WildcardPriority tests
were incorrect because they expected exact-match flows to be overridden by
wildcard flows with higher priority.

Rather than rework the tests to treat exact-match flows as having the highest
priority I've chosen to replace the exact-match flows with wildcard flows. The
reason is that later versions of OpenFlow do not specify this behavior. Once
the test-groups functionality has been added to OFTest we can add a 1.0-only
test that checks the exact-match priority.

Reported-by: Ben Pfaff <blp@nicira.com>