oft: add test-file option
This option reads a file that has one test name or group name per line.
Whitespace and comments starting with '#' are ignored. The resulting tests are
unioned with those given on the command line.

The main usecase for this is freezing a set of tests to be run against a
switch. You can use the --list-test-names option to get the current set of
tests in a format suitable for use with this option, then manually edit
it to add, remove, and comment tests.