Break out double tagging tests
It is not clear in the 1.0 spec what the treatment of double
tagged packets should be.  It doesn't really get clearer in
later specs, though they are more explicitly discussed in 1.1.
Wildcarding VLAN has the effect on some switches of ignoring
all VLAN tags.  This is not a clear violation of the spec.  As
a result, this commit proposes to break out the double tagged
VLAN test cases to a separate test and mark it nonstandard.