tests-1.3: add bsn_flow_checksum testsuite
See the Loxigen extension bsn_flow_checksum. This extension maintains checksums
per table and for a configurable number of buckets in each table. The checksums
are sent to the switch in the cookie field of the flow-add. These checksums are
intended to be used for a hash-tree like synchronization scheme.