dataplane: pass protocol number to bind() instead of socket()
This makes the bind call much faster because it doesn't need to unregister the
protocol hook, which requires waiting for an RCU grace period (~10ms). Linux
3.14 and later optimize this internally, but Ubuntu 14.04 doesn't have a recent
enough kernel.

The Python socket library does the htons for us when creating the sockaddr for
bind.

This change reduces the time to run IVS automated tests by half.