port_stats: check for packets >= expected
If an earlier test run sends a packet through the port, and if the switch takes
a second or so to make those counter updates visible to OpenFlow, then this
test would fail because the port stats would be higher than expected.

Changing the conditional reduces the usefulness of the test because it won't
catch overcounting errors, but it's necessary to make it correct.