tests: add sleep after flow timeouts
For normal flow-mods we can use the barrier message to find out when the change
has made it to the datapath. This isn't the case with timeouts, because the
update isn't associated with any controller connection. When tests are run in
short succession the timed out flow may still exist in the datapath when the
next test starts.