Added a helper function to find out egress port
This test case lb_port as input. This lb_port (provided by the user)
acts as an ingress port at the switch. However the test case tries to
send the packet back to the controller through lb_port+1 which is not
always true. Say I supplied lb_port 102, but 103 may/may not be a part
of an openflow configuration. The fixes shown takes care of this issue.