Added a helper function to find out egress port
This test case takes lb_port as input. This lb_port (provided by the
user) acts as an ingress port at the switch. However the test case tries
to send the packet to the controller through port lb_port+1, which is a
serious bug here, because say 102 is my lb_port, then 103 may/may not be
a part of an openflow configuration. The fixes done by me will take care
of this issue.