Avoid ambiguous case in pktact.DirectLlcPackets test for OF1.0.
OpenFlow 1.0 section 3.4 "Matching" says:

    A flow entry that specifies an Ethernet type of 0x05FF, matches all
    Ethernet 802.2 frames without a SNAP header and those with SNAP headers
    that do not have an OUI of 0x000000.

The two parts of the test changed in this patch are meant to check that
such a match on Ethertype 0x5ff (aka OFP_DL_TYPE_NOT_ETH_TYPE) does not
match a packet with a valid Ethertype that appears in a SNAP header.
However, they in fact test against 0x0003 as the Ethertype, which is not
a valid Ethertype (the minimum valid Ethertype is 0x600).  The OpenFlow
1.0 specification doesn't say anything about how to handle this case.
Open vSwitch chooses to treat it the same as the other non-Ethernet cases
using 0x5ff as the Ethertype.  Thus, the test fails due to an ambiguity.

This commit changes the test to one that uses a valid Ethertype (0x1234).